pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homebase-sync"
version = "0.1.0"
description = "Scrape Homebase weekly schedule and sync shifts to Google Calendar"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Joe Aguirre" }]
license = { text = "Proprietary" }
dependencies = [
    "playwright>=1.49",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "google-api-python-client>=2.150",
    "google-auth>=2.35",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "python-dotenv>=1.0",
    "tzdata>=2024.1; platform_system == 'Windows'",
    "playwright-stealth>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "black>=24.10",
]

[project.scripts]
homebase-sync = "homebase_sync.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.black]
line-length = 100
target-version = ["py311"]
